pyproject.toml
[project]
name = "SLII_openai_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.38.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "pydantic[email]",
    "python-dotenv>=1.1.1",
    "bcrypt==4.0.1",
    "motor>=3.7.1",
    "pymongo>=4.15.1",
    "openai>=2.0.1",
    "black>=25.9.0",
    "openai-agents>=0.0.6",
    "sse-starlette>=3.0.3",
    "beautifulsoup4>=4.14.2",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.2",
    "pydantic[email]>=2.12.4",
    "pytest>=9.0.1",
    "python-docx>=1.2.0",
     "cloudscraper",
     "setuptools>=80.9.0",
    "weasyprint>=67.0",
]
